test: share MiniCPM streaming helpers
